algorithm/pyproject.toml
[tool.poetry]
name = "animalia_project"
version = "0.1.0"
description = "Algorithm components for animalia"
authors = ["shuntakochi <shuntakochi@gmail.com>"]
readme = "README.md"
packages = [
    { include = "common"},
    { include = "recommend_system"},
    { include = "search_engine"},
    { include = "task_scoring_system"},
]

[tool.poetry.dependencies]
python = "^3.12"
ipykernel = "^6.29.5"
geopandas = "^1.0.1"
dask = {extras = ["complete"], version = "^2024.12.1"}
sqlalchemy = "^2.0.37"
jismesh = "^2.1.0"
polars = "^1.19.0"
s3fs = "^2024.12.0"
boto3 = "<1.36.0"
matplotlib = "^3.10.0"
psycopg2-binary = "^2.9.10"
japanize-matplotlib = "^1.1.3"
black = "^24.10.0"
isort = "^6.0.0"
mypy = "^1.14.1"
dask-expr = "1.1.21"
haversine = "^2.9.0"
requests = "^2.32.3"
filelock = "3.18.0"
fsspec = "2024.12.0"
Jinja2 = "3.1.6"
MarkupSafe = "3.0.2"
mpmath = "1.3.0"
networkx = "3.4.2"
numpy = "2.2.3"
nvidia-cublas-cu12 = "12.4.5.8"
nvidia-cuda-cupti-cu12 = "12.4.127"
nvidia-cuda-nvrtc-cu12 = "12.4.127"
nvidia-cuda-runtime-cu12 = "12.4.127"
nvidia-cudnn-cu12 = "9.1.0.70"
nvidia-cufft-cu12 = "11.2.1.3"
nvidia-curand-cu12 = "10.3.5.147"
nvidia-cusolver-cu12 = "11.6.1.9"
nvidia-cusparse-cu12 = "12.3.1.170"
nvidia-cusparselt-cu12 = "0.6.2"
nvidia-nccl-cu12 = "2.21.5"
nvidia-nvjitlink-cu12 = "12.4.127"
nvidia-nvtx-cu12 = "12.4.127"
packaging = "24.2"
pandas = "2.2.3"
protobuf = "6.30.1"
python-dateutil = "2.9.0.post0"
pytz = "2025.1"
setuptools = "76.0.0"
six = "1.17.0"
sympy = "1.13.1"
tensorboardX = "2.6.2.2"
torch = "2.6.0"
tqdm = "4.67.1"
triton = "3.2.0"
typing_extensions = "4.12.2"
tzdata = "2025.1"
supabase = "^2.15.0"
dotenv = "^0.9.9"
uvicorn = "^0.34.0"
fastapi = "^0.115.12"
ftfy = "^6.3.1"
transformers = "^4.50.3"
tiktoken = "^0.9.0"
blobfile = "^3.0.0"
sentencepiece = "^0.2.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"